actions: Split out zerver.actions.create_realm.
