+++ zerver/management/commands/create_realm.py
@@ -3,8 +3,8 @@ from typing import Any
 
 from django.core.management.base import CommandError
 
+from zerver.actions.create_realm import do_create_realm
 from zerver.actions.create_user import do_create_user
-from zerver.lib.actions import do_create_realm
 from zerver.lib.management import ZulipBaseCommand
 from zerver.models import UserProfile
