+++ zerver/tests/test_auth_backends.py
@@ -40,6 +40,7 @@ from social_django.storage import BaseDjangoStorage
 from social_django.strategy import DjangoStrategy
 
 from confirmation.models import Confirmation, create_confirmation_link
+from zerver.actions.create_realm import do_create_realm
 from zerver.actions.create_user import do_create_user, do_reactivate_user
 from zerver.actions.invites import do_invite_users
 from zerver.actions.realm_settings import (
@@ -49,7 +50,6 @@ from zerver.actions.realm_settings import (
 )
 from zerver.actions.user_settings import do_change_password
 from zerver.actions.users import change_user_is_active, do_deactivate_user
-from zerver.lib.actions import do_create_realm
 from zerver.lib.avatar import avatar_url
 from zerver.lib.avatar_hash import user_avatar_path
 from zerver.lib.dev_ldap_directory import generate_dev_ldap_dir
